+++ vclib/src/jvmTest/kotlin/at/asitplus/wallet/lib/jws/JwsServiceJvmTest.kt
@@ -2,17 +2,18 @@ package at.asitplus.wallet.lib.jws
 
 import at.asitplus.crypto.datatypes.CryptoAlgorithm
 import at.asitplus.crypto.datatypes.io.Base64UrlStrict
+import at.asitplus.crypto.datatypes.jws.JweAlgorithm
+import at.asitplus.crypto.datatypes.jws.JweEncrypted
+import at.asitplus.crypto.datatypes.jws.JweEncryption
 import at.asitplus.crypto.datatypes.jws.JwsSigned
 import at.asitplus.wallet.lib.agent.DefaultCryptoService
 import at.asitplus.wallet.lib.data.jsonSerializer
 import com.benasher44.uuid.uuid4
 import com.nimbusds.jose.*
-import com.nimbusds.jose.crypto.ECDSASigner
-import com.nimbusds.jose.crypto.ECDSAVerifier
-import com.nimbusds.jose.crypto.RSASSASigner
-import com.nimbusds.jose.crypto.RSASSAVerifier
+import com.nimbusds.jose.crypto.*
 import io.kotest.assertions.withClue
 import io.kotest.core.spec.style.FreeSpec
+import io.kotest.matchers.nulls.shouldBeNull
 import io.kotest.matchers.nulls.shouldNotBeNull
 import io.kotest.matchers.shouldBe
 import io.matthewnelson.encoding.core.Encoder.Companion.encodeToString
@@ -68,12 +69,26 @@ class JwsServiceJvmTest : FreeSpec({
                 else -> throw IllegalArgumentException("Unknown Key Type") // -||-
             }
 
+            val jweAlgorithm = when (algo) {
+                CryptoAlgorithm.ES256, CryptoAlgorithm.ES384, CryptoAlgorithm.ES512 -> JweAlgorithm.ECDH_ES
+                CryptoAlgorithm.RS256, CryptoAlgorithm.PS256 -> JweAlgorithm.RSA_OAEP_256
+                CryptoAlgorithm.RS384, CryptoAlgorithm.PS384 -> JweAlgorithm.RSA_OAEP_384
+                CryptoAlgorithm.RS512, CryptoAlgorithm.PS512 -> JweAlgorithm.RSA_OAEP_512
+                else -> throw IllegalArgumentException("Unknown JweAlgorithm")
+            }
+
             val jvmVerifier =
-                if (algo.name.take(2) == "ES") ECDSAVerifier(keyPair.public as ECPublicKey)
+                if (algo.isEc) ECDSAVerifier(keyPair.public as ECPublicKey)
                 else RSASSAVerifier(keyPair.public as RSAPublicKey)
             val jvmSigner =
-                if (algo.name.take(2) == "ES") ECDSASigner(keyPair.private as ECPrivateKey)
+                if (algo.isEc) ECDSASigner(keyPair.private as ECPrivateKey)
                 else RSASSASigner(keyPair.private as RSAPrivateKey)
+            val jvmEncrypter =
+                if (algo.isEc) ECDHEncrypter(keyPair.public as ECPublicKey)
+                else RSAEncrypter(keyPair.public as RSAPublicKey)
+            val jvmDecrypter =
+                if (algo.isEc) ECDHDecrypter(keyPair.private as ECPrivateKey)
+                else RSADecrypter(keyPair.private as RSAPrivateKey)
 
             val cryptoService = DefaultCryptoService(keyPair, algo)
             val jwsService = DefaultJwsService(cryptoService)
@@ -90,7 +105,7 @@ class JwsServiceJvmTest : FreeSpec({
                         jwsService.createSignedJwt(JwsContentTypeConstants.JWT, stringPayload.encodeToByteArray())
                     signed.shouldNotBeNull()
                     val selfVerify = verifierJwsService.verifyJwsObject(signed)
-                    withClue("$algo: Signature: ${signed.signature.serialize()}"){
+                    withClue("$algo: Signature: ${signed.signature.serialize()}") {
                         selfVerify shouldBe true
                     }
                 }
@@ -140,115 +155,53 @@ class JwsServiceJvmTest : FreeSpec({
                         result shouldBe true
                     }
                 }
+
+                /**
+                 * Encryption is currently only supported for EC-Keys see issue `https://github.com/a-sit-plus/kmm-vc-library/issues/29`
+                 */
+                if(thisConfiguration.first == "EC") {
+                    "Encrypted object from ext. library can be decrypted with int. library" {
+                        val stringPayload = jsonSerializer.encodeToString(randomPayload)
+                        val libJweHeader = JWEHeader.Builder(JWEAlgorithm(jweAlgorithm.text), EncryptionMethod.A256GCM)
+                            .type(JOSEObjectType(JwsContentTypeConstants.DIDCOMM_ENCRYPTED_JSON))
+                            .contentType(JwsContentTypeConstants.DIDCOMM_PLAIN_JSON)
+                            .keyID(cryptoService.jsonWebKey.keyId)
+                            .build()
+                        val libJweObject = JWEObject(libJweHeader, Payload(stringPayload)).also {
+                            it.encrypt(jvmEncrypter)
+                        }
+                        val encryptedJwe = libJweObject.serialize()
+
+                        val parsedJwe = JweEncrypted.parse(encryptedJwe)
+                        parsedJwe.shouldNotBeNull()
+
+                        val result = jwsService.decryptJweObject(
+                            parsedJwe, encryptedJwe
+                        )
+
+                        result?.payload?.decodeToString() shouldBe stringPayload
+                    }
+
+                    "Encrypted object from int. library can be decrypted with ext. library" {
+                        val stringPayload = jsonSerializer.encodeToString(randomPayload)
+                        val encrypted = jwsService.encryptJweObject(
+                            JwsContentTypeConstants.DIDCOMM_ENCRYPTED_JSON,
+                            stringPayload.encodeToByteArray(),
+                            cryptoService.jsonWebKey,
+                            JwsContentTypeConstants.DIDCOMM_PLAIN_JSON,
+                            jweAlgorithm,
+                            JweEncryption.A256GCM,
+                        )
+
+                        val parsed = JWEObject.parse(encrypted)
+                        parsed.shouldNotBeNull()
+                        parsed.payload.shouldBeNull()
+
+                        parsed.decrypt(jvmDecrypter)
+                        parsed.payload.toString() shouldBe stringPayload
+                    }
+                }
             }
         }
     }
-//
-//    "EC" - {
-//        withData(256, 384, 521)
-//        { bits ->
-//            publicKeyDesc = List<Pair<KeyPair, CryptoAlgorithm>>(5) {
-//                val keyPair = KeyPairGenerator.getInstance("EC").apply {
-//                    initialize(bits)
-//                }.genKeyPair()
-//                val algo = when (bits) {
-//                    256 -> CryptoAlgorithm.ES256
-//                    384 -> CryptoAlgorithm.ES384
-//                    521 -> CryptoAlgorithm.ES512
-//                    else -> CryptoAlgorithm.RS1.also { throw IllegalArgumentException("Unknown EC Curve size") } // necessary(compiler), but otherwise redundant else-branch
-//                }
-//                Pair(keyPair, algo)
-//            }
-//            withData(publicKeyDesc) { (keyPair, algorithm) ->
-//                cryptoService = DefaultCryptoService(keyPair, algorithm)
-//                jwsService = DefaultJwsService(cryptoService)
-//                verifierJwsService = DefaultVerifierJwsService()
-//                randomPayload = uuid4().toString()
-//
-//                withData("signed object from ext. library can be verified: $algorithm", Pair(keyPair, algorithm)) {
-//                    val stringPayload = jsonSerializer.encodeToString(randomPayload)
-//                    val libHeader = JWSHeader.Builder(JWSAlgorithm(algorithm.name)).type(JOSEObjectType("JWT"))
-//                        .keyID(cryptoService.jsonWebKey.keyId).build()
-//                    val libObject = JWSObject(libHeader, Payload(stringPayload)).also {
-//                        it.sign(ECDSASigner(keyPair.private as ECPrivateKey))
-//                    }
-//                    libObject.verify(ECDSAVerifier(keyPair.public as ECPublicKey)) shouldBe true
-//
-//                    // Parsing to our structure verifying payload
-//                    val signedLibObject = libObject.serialize()
-//                    val parsedJwsSigned = JwsSigned.parse(signedLibObject)
-//                    parsedJwsSigned.shouldNotBeNull()
-//                    parsedJwsSigned.payload.decodeToString() shouldBe stringPayload
-//                    val parsedSig = parsedJwsSigned.signature.rawByteArray.encodeToString(Base64UrlStrict)
-//
-//                    withClue(
-//                        "Signatures should match\n" +
-//                                "Ours:\n" +
-//                                "$parsedSig\n" +
-//                                "Theirs:\n" +
-//                                "${libObject.signature}"
-//                    ) {
-//                        parsedSig shouldBe libObject.signature.toString()
-//                    }
-//
-//                    // verifying external JWT with our service
-//                    withClue("Signature is valid") {
-//                        val result = verifierJwsService.verifyJwsObject(parsedJwsSigned)
-//                        result shouldBe true
-//                    }
-//                }
-//
-//                withData("signed object can be verified with ext. library: $algorithm", Pair(keyPair, algorithm)) {
-//                    val stringPayload = jsonSerializer.encodeToString(randomPayload)
-//                    val signed =
-//                        jwsService.createSignedJwt(JwsContentTypeConstants.JWT, stringPayload.encodeToByteArray())
-//
-//                    verifierJwsService.verifyJwsObject(signed!!) shouldBe true
-//                    val parsed = JWSObject.parse(signed.serialize())
-//                    parsed.shouldNotBeNull()
-//                    parsed.payload.toString() shouldBe stringPayload
-//                    val result = parsed.verify(ECDSAVerifier(keyPair.public as ECPublicKey))
-//                    result shouldBe true
-//                }
-//
-//                withData("encrypted object from ext. library can be decrypted: $algorithm", Pair(keyPair, algorithm)) {
-//                    val stringPayload = jsonSerializer.encodeToString(randomPayload)
-//                    val libJweHeader = JWEHeader.Builder(JWEAlgorithm.ECDH_ES, EncryptionMethod.A256GCM)
-//                        .type(JOSEObjectType(JwsContentTypeConstants.DIDCOMM_ENCRYPTED_JSON))
-//                        .contentType(JwsContentTypeConstants.DIDCOMM_PLAIN_JSON).keyID(cryptoService.jsonWebKey.keyId)
-//                        .build()
-//                    val libJweObject = JWEObject(libJweHeader, Payload(stringPayload)).also {
-//                        it.encrypt(ECDHEncrypter(keyPair.public as ECPublicKey))
-//                    }
-//                    val encryptedJwe = libJweObject.serialize()
-//
-//                    val parsedJwe = JweEncrypted.parse(encryptedJwe)
-//                    parsedJwe.shouldNotBeNull()
-//
-//                    jwsService.decryptJweObject(
-//                        parsedJwe, encryptedJwe
-//                    )?.payload?.decodeToString() shouldBe stringPayload
-//                }
-//
-//                withData("encrypted object can be decrypted with ext. library: $algorithm", Pair(keyPair, algorithm)) {
-//                    val stringPayload = jsonSerializer.encodeToString(randomPayload)
-//                    val encrypted = jwsService.encryptJweObject(
-//                        JwsContentTypeConstants.DIDCOMM_ENCRYPTED_JSON,
-//                        stringPayload.encodeToByteArray(),
-//                        cryptoService.jsonWebKey,
-//                        JwsContentTypeConstants.DIDCOMM_PLAIN_JSON,
-//                        JweAlgorithm.ECDH_ES,
-//                        JweEncryption.A256GCM,
-//                    )
-//
-//                    val parsed = JWEObject.parse(encrypted)
-//                    parsed.shouldNotBeNull()
-//                    parsed.payload.shouldBeNull()
-//
-//                    parsed.decrypt(ECDHDecrypter(keyPair.private as ECPrivateKey))
-//                    parsed.payload.toString() shouldBe stringPayload
-//                }
-//            }
-//        }
-//    }
 })